backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.4,<5
fastapi>=0.115,<1
python-multipart>=0.0.9,<1
httpx>=0.27,<1
uvicorn[standard]>=0.32,<1
sqlalchemy>=2.0,<3
psycopg[binary]>=3.2,<4
pydantic-settings>=2.0,<3
python-dotenv>=1.0,<2
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7.4,<2
bcrypt>=4.0.1,<4.1
email-validator>=2.0,<3
